% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_in_white_noise.R
\name{acf_in_white_noise}
\alias{acf_in_white_noise}
\title{Compute confidence intervals around an autocorrelation function (ACF)}
\usage{
acf_in_white_noise(acf, ci = 0.95, return = "upper")
}
\arguments{
\item{acf}{An ACF object, produced by \code{\link[stats]{acf}}.}

\item{ci}{An numeric value between 0 and 1 which defines the confidence interval (CI) to be calculated. The default is 0.95; i.e, a 95 percent CI.}

\item{return}{A character input specifying \code{"upper"} or \code{"CI"}. If \code{return = "upper"}, a single number is returned which is the upper 95 percent confidence limit of the extent of autocorrelation in a sample of white noise. If \code{return = "CI"} a list comprising \code{CI$fit (0)}, \code{CI$lowerCI} (the lower CI) and \code{CI$upperCI} (the upper CI) is returned. This can be plotted by \code{\link[plot.pretty]{add_model_predictions}}.}
}
\value{
The function returns a number or a list, depending on the input to \code{return} (see above).
}
\description{
This function calculates confidence intervals around an ACF. This represents the upper limit of autocorrelation in white noise. This is useful for producing pretty plots.
}
\seealso{
\code{\link[stats]{acf}}, \code{\link[plot.pretty]{add_model_predictions}}
}
\author{
Edward Lavender
}
