% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_ts.R
\name{average_ts}
\alias{average_ts}
\title{Summarise a time series in bins}
\usage{
average_ts(
  dat,
  split = NULL,
  timestamp,
  response,
  breaks,
  fun = mean,
  first = NULL,
  average = NULL,
  as_time = as.POSIXct,
  order = c(timestamp, response, first, average),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dat}{A dataframe which includes observations collected over time.}

\item{split}{(optional) A character that defines the name of the column in \code{dat} that distinguishes independent time series.}

\item{timestamp}{A character that defines the name of the column in \code{dat} that contains timestamps, to be grouped into bins.}

\item{response}{A character that defines the name of the column in \code{dat} that contains the values of the response to be averaged in each bin.}

\item{breaks}{A number or vector which defines the number of unique cut points or the unique cut points themselves at which each independent time series is cut into bins. This is passed to \code{\link[base]{cut}}.}

\item{fun}{A function used to calculate the average response and the average value of any other other numeric variable specified in \code{average} in each bin.}

\item{first}{(optional) A character vector that defines the name(s) of the column(s) in \code{dat} for which the first observation for each independent time series should be retained in the averaged dataframe. This is useful for 'static' variables which do not change through time (e.g., the sex of each individual).}

\item{average}{(optional) A character vector that defines the name(s) of any other column(s) in \code{dat} that need to be averaged in each bin, like \code{response}.}

\item{as_time}{A function that converts a character vector of breaks into a vector of times. The default function is \code{as.POSIXct}.}

\item{order}{A character vector of column names that defines the order of desired columns in the returned dataframe.}

\item{verbose}{A logical input that defines whether or not to return messages to the console to monitor function progress.}

\item{...}{Additional arguments passed to \code{\link[base]{cut}}.}
}
\value{
The function returns a dataframe in which timestamps have been aggregated into a sequence of bins and the value of the response and any other numeric columns requested have been averaged over all the data in each bin. For static variables (e.g., the sex of an individual), the first observation for each independent time series may also be included.
}
\description{
This function averages a time series (i.e., all of the values of a response variable collected through time) in each of a series of user-specified bins. The function can handle independent time series (e.g., for multiple individuals), average numeric explanatory variables alongside the response, incorporate other 'static' variables (e.g., the sex of each individual) and implement different types of averages.
}
\examples{
dat_flapper_av <- average_ts(dat = dat_flapper,
                             split = "id",
                             timestamp = "timestamp",
                             response = "depth",
                             breaks = "hours",
                             first = "id",
                             average = "temp"
                             )
utils::head(dat_flapper_av)
}
\seealso{
\code{\link[prettyGraphics]{summarise_in_bins}}
}
\author{
Edward Lavender
}
