% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_dates.R
\name{cum_dates}
\alias{cum_dates}
\title{The total number of events through time (in days)}
\usage{
cum_dates(data, ...)
}
\arguments{
\item{data}{A vector of dates.}

\item{...}{Other arguments passed to \code{\link[base]{as.Date}}.}
}
\value{
The function returns a dataframe with three columns: (1) 'date', the date; (2) 'freq', the number of observations on each unique date; and (3) 'cumul', the total number of events by each date (i.e. the cumulative frequency).
}
\description{
Calculate the total number of events by each date in a series of dates (i.e., a cumulative frequency distribution).
}
\examples{

#### Example (1): A simple sequence of dates
cum_dates(seq.Date(as.Date("2016-01-01"), as.Date("2016-01-10"), 1))

#### Example (2): Multiple records of each date
set.seed(1)
dates <- seq.Date(as.Date("2016-01-01"), as.Date("2016-01-10"), 1)
dates <- sort(sample(dates, size = 1000, replace = TRUE))
cum_dates(dates)

}
\author{
Edward Lavender
}
