% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates_of_change.R
\name{difference}
\alias{difference}
\title{Calculate the difference between two numbers or times}
\usage{
difference(x2, x1, f = NULL, ...)
}
\arguments{
\item{x2}{A number, \code{\link[base]{Dates}} or \code{\link[base]{DateTimeClasses}} object.}

\item{x1}{A number, \code{\link[base]{Dates}} or \code{\link[base]{DateTimeClasses}} object.}

\item{f}{A function to process outputs.}

\item{...}{Other arguments passed to \code{\link[base]{difftime}} e.g. \code{units = "secs"}.}
}
\value{
A difference.
}
\description{
Calculate a difference between two numbers or times. For numeric objects, the difference is calculated as \code{x2 - x1}. For time objects (i.e., \code{\link[base]{DateTimeClasses}} or \code{\link[base]{Dates}}), the difference is calculated using \code{\link[base]{difftime}}. A function can be supplied to process outputs; e.g. \code{\link[base]{abs}}, or \code{\link[base]{numeric}} for \code{\link[base]{difftime}} objects.
}
\examples{

#### Example (1) Numeric difference
difference(1, 2)

#### Example (2) Numeric difference with post processing
difference(1, 2, f = abs)

#### Example (3) Difference with times
difference(as.POSIXct("2016-01-05"), as.POSIXct("2016-01-01"))

#### Example (4) Difference with times and units specified
difference(as.POSIXct("2016-01-05"), as.POSIXct("2016-01-01"), units = "mins")

#### Example (5) Difference with times, specified units and post-processing
difference(as.POSIXct("2016-01-05"), as.POSIXct("2016-01-01"), units = "mins", f = as.numeric)

}
\author{
Edward Lavender
}
