% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_ts.R
\name{flag_ts}
\alias{flag_ts}
\title{Flag independent segments of timeseries}
\usage{
flag_ts(x, fct = NULL, dat = NULL, duration_threshold, flag = 1:3)
}
\arguments{
\item{x}{A vector of timestamps, either as an integer/numeric object, a \code{\link[base]{DateTimeClasses}} or a \code{\link[base]{Date}}.}

\item{fct}{A grouping factor defining independent segments of timeseries (e.g. individuals).}

\item{dat}{(optional) A dataframe with a column \code{x} and, optionally, a column \code{fct}, can be supplied instead of \code{x} and \code{fct}.}

\item{duration_threshold}{A numeric input defining the number of units (if \code{x} is an integer/numeric) or the number of minutes (if \code{x} is \code{\link[base]{DateTimeClasses}} or a \code{\link[base]{Date}}) between records after which separated timeseries are considered effectively 'independent'. This could be informed by a the autocorrelation function of the residuals of a model without any autocorrelation.}

\item{flag}{A numeric input/vector of \code{1}, \code{2} or \code{3} specifying the flag type to be returned. \code{flag = 1} returns a logical vector with \code{TRUE} marking the first observation in each independent segment of timeseries (e.g. as required by \code{\link[mgcv]{bam}}. \code{flag = 2} returns identifies the first, second,... nth, independent segment of each factor level's timeseries. \code{flag = 3} provides a unique identifier for each segment of timeseries (e.g. as required by \code{\link[mgcv]{gamm}}).}
}
\value{
A dataframe with the duration between each observation and the next observation and corresponding flags.
}
\description{
The function 'flags' independent sections of timeseries using up to three methods. This is useful for identifying gaps in timeseries and/or for models of timeseries data in which independent segments of timeseries need to be treated as such. To flag timeseries, the function can consider two drivers of independence: (a) a grouping factor (\code{fct}) which defines inherently independent timeseries in a dataset (e.g. a dataset may comprise timeseries for different individuals) and (b) gaps in the timeseries which, when greater than a user-defined threshold (\code{duration_threshold}), separate timeseries that can be considered effectively independent, even if they are derived from the same level of a grouping factor. Using these two criteria, timeseries can be flagged using three methods which different modelling approaches may require.
}
\details{
Inputs should be ordered by ordered by \code{fct} (if applicable) then \code{x}.
}
\examples{

#### Define some irregularly spaced timeseries
t <- c(seq.POSIXt(as.POSIXct("2016-01-01"), as.POSIXct("2016-01-02"), by = "6 hours"),
       as.POSIXct("2016-01-02 12:00:00"),
       seq.POSIXt(as.POSIXct("2016-01-02 18:00:00"), as.POSIXct("2016-01-03"), by = "5 hours")
)

#### Example (1) Supply a vector of timestamps to flag independent sections of timeseries
flag_ts(
  x = t,
  duration_threshold = 8*60,
  flag = 1:3)

#### Example (2) Supply a dataframe with a timestamp column instead
flag_ts(dat = data.frame(x = t),
        duration_threshold = 8*60,
        flag = 1:3)

#### Example (3) Supply a factor level which separates unique timeseries
fct_levels <- c(rep(1, length(t)), rep(2, length(t)))
t2 <- rep(t, 2)
flag_ts(
  x = t2,
  fct = fct_levels,
  duration_threshold = 8*60,
  flag = 1:3)

#### Example (4) Supply timeseries and a factor in a dataframe organised by fct then timestamp
flag_ts(dat = data.frame(x = t2, fct = fct_levels),
        duration_threshold = 8*60,
        flag = 1:3)

#### Example (5) Numeric example without factor
x <- c(seq(1, 5, by = 1), seq(100, 105, by = 1))
flag_ts(x = x,
        fct = NULL,
        dat = NULL,
        duration_threshold = 5,
        flag = 1:3)

#### Example (6) Numeric example with factor
fct_levels <- c(rep(1, length(x)), rep(2, length(x)))
x <- rep(x, 2)
flag_ts(x = x,
        fct = fct_levels,
        dat = NULL,
        duration_threshold = 5,
        flag = 1:3)

#### Example (7) Numeric example with factor via dat argument
cbind(x,
      flag_ts(dat = data.frame(x = x, fct = fct_levels),
              duration_threshold = 5,
              flag = 1:3)
)

}
\author{
Edward Lavender
}
