% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gams4dts_bam_model.R, R/gams4dts_support.R
\name{gams4dts_bam_model}
\alias{gams4dts_bam_model}
\title{Evaluate an \code{mgcv::bam()} model for GAMS4DTS}
\usage{
gams4dts_bam_model(
  f,
  likelihood = list(distribution = "gaussian", link = "identity"),
  rho = 0,
  knots,
  gamma = 1,
  dat
)

gams4dts_bam_model(
  f,
  likelihood = list(distribution = "gaussian", link = "identity"),
  rho = 0,
  knots,
  gamma = 1,
  dat
)
}
\arguments{
\item{f}{A formula (see \code{\link[mgcv]{bam}}).}

\item{likelihood}{A list specifying the likelihood distribution and link function.}

\item{rho}{rho (see \code{\link[mgcv]{bam}}).}

\item{knots}{knots (see \code{\link[mgcv]{bam}}).}

\item{gamma}{gamma (see \code{\link[mgcv]{bam}}).}

\item{dat}{A dataframe containing the data used to fit the model (see \code{\link[mgcv]{bam}}).}
}
\value{
An \code{\link[mgcv]{bam}} model object.

An \code{\link[mgcv]{bam}} model object.
}
\description{
This function is a wrapper for \code{\link[mgcv]{bam}} which passes user-defined inputs from \code{\link[Tools4ETS]{GAMS4DTS}} to \code{\link[mgcv]{bam}}. The function is not intended to be called outside of \code{\link[Tools4ETS]{GAMS4DTS}}.

This function is a wrapper for \code{\link[mgcv]{bam}} which passes user-defined inputs from \code{\link[Tools4ETS]{GAMS4DTS}} to \code{\link[mgcv]{bam}}. The function is not intended to be called outside of \code{\link[Tools4ETS]{GAMS4DTS}}.
}
\author{
Edward Lavender

Edward Lavender
}
