% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gams4dts_support.R
\name{gams4dts_define_bam_formula}
\alias{gams4dts_define_bam_formula}
\title{Define an \code{\link[mgcv]{bam}} formula in \code{\link[Tools4ETS]{GAMS4DTS}}}
\usage{
gams4dts_define_bam_formula(
  response = "depth",
  fixed_covariates,
  smooth_covariates,
  dat
)
}
\arguments{
\item{response}{A character input specifying the name of the response variable in \code{dat} (i.e. \code{"depth"}).}

\item{fixed_covariates}{A list containing covariates implemented as fixed effects.}

\item{smooth_covariates}{A list containing smooth covariates, including their names and basis functions.}

\item{dat}{A dataframe.}
}
\value{
An \code{\link[mgcv]{bam}} model formula for \code{\link[Tools4ETS]{GAMS4DTS}}.
}
\description{
This function is used to pass user-defined variables in \code{\link[Tools4ETS]{GAMS4DTS}} into an \code{\link[mgcv]{bam}} formula. The function is not intended to be called outside of \code{\link[Tools4ETS]{GAMS4DTS}}.
}
\author{
Edward Lavender
}
\keyword{internal}
