% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gams4dts_support.R
\name{gams4dts_thin_ts}
\alias{gams4dts_thin_ts}
\title{\code{\link[Tools4ETS]{GAMS4DTS}} Thinning Methods}
\usage{
gams4dts_thin_ts(
  dat,
  id_column = "individual",
  timestamp_column = "timestamp",
  start_event_id_column = "start_event_id",
  start_event_logic_column = "start_event",
  method = c("swa", "sma", "sps"),
  parameters = list(swa = list(time_mins = 60, latitude = 56.41041, longitude =
    5.469723, covariates = c("sex", "length", "sun_angle", "tidal_height", "lunar_phase",
    "julian_day"), independence_threshold = 720), sps = list(first = c(1, 2), nth = 10),
    sma = list())
)
}
\arguments{
\item{dat}{A dataframe to be thinned.}

\item{id_column}{A character input defining the column name in \code{dat} which distinguishes individuals (i.e. \code{"individual"}).}

\item{timestamp_column}{A character input defining the column name in \code{dat} which contains timestamps (i.e. \code{"timestamp"}).}

\item{start_event_id_column}{A character input defining the column name in \code{dat} which identifies start events (i.e. \code{"start_event_id"}).}

\item{start_event_logic_column}{A character input defining the column name in \code{dat} which defines start events with TRUE/FALSE (i.e. \code{"start_event"})}

\item{method}{A character input defining the thinning method to be implemented. Supported options are: \code{"swa"}, \code{"sma"} or \code{"sps"}.}

\item{parameters}{A list of parameters that are required to implement thinning depending on the method.}
}
\value{
A thinned dataframe.
}
\description{
This function implements thinning behind the scenes for \code{\link[Tools4ETS]{GAMS4DTS}}. The function is not intended to be called outside of this environment.
}
\author{
Edward Lavender
}
\keyword{internal}
