% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_shiny.R
\name{launch_shiny}
\alias{launch_shiny}
\title{Launch one of Tools4ETS's Shiny applications}
\usage{
launch_shiny(app, data = NULL)
}
\arguments{
\item{app}{A character vector defining the name of the application to be launched. The currently supported options are: \code{"visTS"}, which calls \code{\link[Tools4ETS]{visTS}}; \code{"GAMS4DTS"}, which calls \code{\link[Tools4ETS]{GAMS4DTS}}.}

\item{data}{A dataframe passed to \code{\link[Tools4ETS]{visTS}}. As a minimum, this should includes a column defining timestamps.}
}
\value{
An interactive Shiny-Dashboard interface.
}
\description{
A wrapper function for Shiny applications which enables these to be launched using a common framework.
}
\examples{

#### Example (1) visTS
\dontrun{
launch_shiny("visTS", data = dat_flapper)
}

#### Example (2) GAMS4DTS
\dontrun{
launch_shiny("GAMS4DTS")
}

}
\author{
Edward Lavender
}
