% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_mathematical_functions.R
\name{linear}
\alias{linear}
\title{A linear function}
\usage{
linear(x, a, b)
}
\arguments{
\item{x}{A numeric vector which defines the values at which the linear function will be evaluated.}

\item{a}{A numeric input which defines the intercept.}

\item{b}{A numeric input which defines the gradient.}
}
\value{
A number or a numeric vector.
}
\description{
This function evaluates a linear function.
}
\examples{
linear(1:10, 5, 3)

}
\author{
Edward Lavender
}
