% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterise.R
\name{parameterise_contrast_2l}
\alias{parameterise_contrast_2l}
\title{Define and visualise a contrast between two factor levels}
\usage{
parameterise_contrast_2l(
  x = factor(c("F", "M")),
  param = 0,
  plot_gam = FALSE,
  dat,
  model,
  term = "sex",
  cex.axis = 1,
  mtext_args = list(list(side = 1, text = "Sex (F, female; M, male)", line = 2.5),
    list(side = 2, text = "Contrast (m)", line = 2.5))
)
}
\arguments{
\item{x}{A vector of two factor levels for which to define a contrast.}

\item{param}{A numeric input which defines the size of the contrast between two factor levels.}

\item{plot_gam}{A logical input which defines whether or not to add model estimates to the plot.}

\item{dat}{The dataframe used to fit the model.}

\item{model}{The model.}

\item{term}{A character input which defines the name of the term in the model.}

\item{cex.axis}{A numeric input which defines the size of the font for axis tick labels.}

\item{mtext_args}{A named list of arguments which are passed to \code{\link[graphics]{mtext}} to produce axis labels. A nested list can be used to control each axis (see also \code{\link[prettyGraphics]{pretty_ts}}).}
}
\value{
The function returns a plot showing the size of the contrast between factor levels, possibly in relation to that estimated by an \code{mgcv} GAM model.
}
\description{
This function is used to define and visualise a contrast between two factor levels (e.g. the difference in depth between males and females). Simulated contrasts can be compared to estimated contrasts from an \code{mgcv} generalised additive model (GAM). The function is intended primarily for use behind-the-scenes in \code{\link[Tools4ETS]{GAMS4DTS}}.
}
\author{
Edward Lavender
}
