% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_mathematical_functions.R
\name{quadratic}
\alias{quadratic}
\title{A quadratic function}
\usage{
quadratic(x, a, b, h, k)
}
\arguments{
\item{x}{A numeric vector which defines the values at which the quadratic function will be evaluated.}

\item{a}{A numeric value which defines the vertical stretch of the function.}

\item{b}{A numeric value which defines the horizontal stretch of the function.}

\item{h}{The x coordinate of the vertex.}

\item{k}{The y coordinate of the vertex.}
}
\value{
A number or a numeric vector.
}
\description{
This function evaluates a quadratic function.
}
\examples{
quadratic(x = 1:10, a = 1, b = 1, h = 0.5, k = 0.25)

}
\author{
Edward Lavender
}
