% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_mathematical_functions.R
\name{sigmoid}
\alias{sigmoid}
\title{A sigmoidal function}
\usage{
sigmoid(x, x0, L, k)
}
\arguments{
\item{x}{A numeric vector which defines the values at which the sigmoidal function will be evaluated.}

\item{x0}{A numeric value which defines the x-value of the sigmoid's midpoint}

\item{L}{A numeric value which defines expected change in f(x) above/below 0. For example, if \code{L = 2}, then the function will stretch ± 1 units above/below 0 at f(0) = 0.}

\item{k}{A numeric value which affects the steepness of the function.}
}
\value{
A number or a numeric vector.
}
\description{
This function evaluates a sigmoidal function.
}
\examples{

x <- 1:10
plot(x, sigmoid(x, 5, L = 5, k = 1), type = "b")
abline(h = 0, lty = 3)

x <- 0:20
plot(x, sigmoid(x, 5, L = 2, k = 1), type = "b")
abline(h = 0, lty = 3)

x <- -60:60
plot(x, sigmoid(x, 0, L = 2, k = 0.1), type = "b")
abline(h = 0, lty = 3)

}
\author{
Edward Lavender
}
