% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sinusoid.R
\name{sim_sinusoid}
\alias{sim_sinusoid}
\title{Simulate and visualise sinusoidal waves with different parameters}
\usage{
sim_sinusoid(
  x,
  f = sin,
  a = 1,
  b = NULL,
  period = 2 * pi,
  c = 0,
  d = 0,
  plot = TRUE,
  type = "l",
  return = FALSE,
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{f}{The sinusoidal function (i.e., sin or cos, see \code{\link[base]{Trig}}).}

\item{a}{A number which defines the amplitude (i.e., the height) of the wave.}

\item{b}{A number which defines the horizontal stretch parameter. Alternatively, \code{period} can be provided.}

\item{period}{A number which defines the period (i.e., the number of units between successive waves) of the wave. This is silently ignored if \code{b} is provided.}

\item{c}{A number which defines the horizontal shift parameter.}

\item{d}{A number which defines the vertical shift parameter (i.e., the mean).}

\item{plot}{A logical input which defines whether or not to plot the simulated wave. This is \code{TRUE} by default.}

\item{type}{A character which defines the type of plot to produce (see \code{\link[graphics]{plot.default}}). A line plot (\code{type = "l"}) is the default.}

\item{return}{A logical input which defines whether or not to return the values of the wave. This is \code{FALSE} by default.}

\item{...}{Other arguments passed to \code{\link[prettyGraphics]{pretty_plot}} to customise the plot.}
}
\description{
This function is used to simulate and visualise sinusoidal waves with different parameters.
}
\details{
Sinusoidal waves are simulated according to the equation \eqn{a f(bx + c) + d} where the parameters are as defined above and \code{b} is related to \code{period} by \eqn{b = \frac{2 \pi}{period}}.
}
\examples{
# Simulate some x values with observations every 0.1 units (e.g., 0.1 s)
x <- seq(0, 10, by = 0.1)
# Visualise a sine wave with a period of 2 s
sim_sinusoid(x, period = 10)
# Or, equivalently, a wave with b = (2*pi)/10
sim_sinusoid(x, b = (2*pi)/10)
# Adjust other parameters, such as the amplitude:
sim_sinusoid(x, period = 10, a = 10)
# Adjust the sinusoidal function via f
sim_sinusoid(x, f = cos, period = 10, a = 10)
}
\author{
Edward Lavender
}
