% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_categories.R
\name{yday2date}
\alias{yday2date}
\title{Convert a Julian day to a date/month/season}
\usage{
yday2date(yday, origin = "2016-01-01", verbose = TRUE, return_list = FALSE)
}
\arguments{
\item{yday}{A numeric input specifying the Julian day (the number of days since January 1st).}

\item{origin}{A character date (YYYY-MM-DD) that defines the first day of the year from which Julian day has been calculated.}

\item{verbose}{A logical input that defines whether or not to print the date, month and season.}

\item{return_list}{A logical input that defines whether or not to return the date, month  and year in a list.}
}
\description{
This function converts a Julian day to a date/month/season in any given year. This is useful, for instance, when making inferences from models of a response ~ Julian day, if you want to quickly draw inferences from models in terms of time units that are more familiar.
}
\examples{
yday2date(16)
yday2date(16, return_list = FALSE)
yday2date(16, verbose = FALSE)

}
\author{
Edward Lavender
}
