% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_categories.R
\name{yday_dbl}
\alias{yday_dbl}
\title{Compute the fraction of the year (days)}
\usage{
yday_dbl(time)
}
\arguments{
\item{time}{A vector of time(s) of class \code{\link[base]{Date}} or \code{\link[base]{DateTimeClasses}}.}
}
\value{
The function returns a number for each element in \code{time} which is the Julian day as a fraction of the total number of days in that year.
}
\description{
This function computes the Julian day on inputted times/dates as a fraction of the total number of days in each year.
}
\examples{
# Compare the start of a non-leap and leap year:
yday_dbl(as.Date("2015-01-01"))
yday_dbl(as.Date("2016-01-01"))
# The end of the year
yday_dbl(as.Date("2015-12-31"))
yday_dbl(as.Date("2015-12-31"))
}
\author{
Edward Lavender
}
