% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gams4dts_support.R
\name{barplot_ts_sex}
\alias{barplot_ts_sex}
\title{Create a barplot of the number of individuals of each sex in time series}
\usage{
barplot_ts_sex(
  dat,
  id_column = "individual",
  f_column = "sex",
  cex.axis = 1,
  cex = 1
)
}
\arguments{
\item{dat}{A dataframe which contains a column which defines unique individuals and a column which defines their sex. Multiple observations from the same individual are supported because the function obtains the sex of each unique individual once.}

\item{id_column}{A character input which defines the name of the column in \code{dat} which distinguishes individuals.}

\item{f_column}{A character input which defines the name of the column in \code{dat} which defines the sex of each individual.}

\item{cex.axis}{A number which defines the font size for axis tick labels.}

\item{cex}{A number which defines the font sizes for axis labels.}
}
\value{
The function returns a pre-customised barplot showing the number of individuals of each sex in a dataframe.
}
\description{
This function produces a barplot showing the number of individuals of each sex in a time series. To do this, the function takes in a dataframe which contains a column distinguishing individuals and a column defining their sex, which are specified as character strings. The function then identifies the sex of each unique individual, sums the number of individuals of each sex and produces a barplot. This function is primarily intended for use in \code{\link[Tools4ETS]{GAMS4DTS}}.
}
\author{
Edward Lavender
}
\keyword{internal}
