% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMS4DTS.R
\name{GAMS4DTS}
\alias{GAMS4DTS}
\title{GAMS4DTS}
\usage{
GAMS4DTS()
}
\description{
An interactive Shiny-Dashboard interface for exploring generalised additive models (GAMs) as a tool for modelling depth timeseries (DTS). A specific vignette provides further information for this function.
}
\examples{
\dontrun{
GAMS4DTS()
}

}
\author{
Edward Lavender
}
