% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_recapture.R
\name{suggest_recapture}
\alias{suggest_recapture}
\title{Explore potential, unrecorded recapture events within depth timeseries}
\usage{
suggest_recapture(
  data,
  threshold_depth = 1,
  plot = TRUE,
  window = 60 * 60 * 12,
  xlab = "Timestamp",
  ylab = "Depth",
  prompt = TRUE,
  ndates_warning = 100,
  ...
)
}
\arguments{
\item{data}{A dataframe with two named columns: 'timestamp' (in POSIXct format) and 'depth' (of a given individual).}

\item{threshold_depth}{A numeric value which defines the depth threshold; i.e. the depth such that when the animal is at or above (shallower) than this depth, a recapture event may have occurred and should be checked.}

\item{plot}{A logical input which defines whether or not to create a plot.}

\item{window}{A numeric value which defines the number of seconds either side of first and last potential 'recapture' event on any given day for which the depth timeseries is plotted. The default is 43200 s (i.e. 12 hours).}

\item{xlab}{A label for the x axis.}

\item{ylab}{A label for the y axis.}

\item{prompt}{A logical value which defines whether or not to pause following the display of each graph. If TRUE, the user needs to press Enter prior to moving on to the next recapture event.}

\item{ndates_warning}{A numeric value defining the number of dates with putative recapture events above which the function will warn the user to reconsider \code{prompt = TRUE}.}

\item{...}{Extra arguments passed to \code{\link[graphics]{plot}}, excluding \code{xlim} which is set based on the window argument provided.}
}
\value{
A dataframe with 4 columns: 'timestamp' (as inputted), 'depth' (as inputted), 'recapture_window' (a logical value specifying whether the position is inside a recapture window, defined by \code{window}), and 'above_depth_threshold' (a logical value which specifies the positions at which the depth is shallower than the depth threshold). A plot can also be returned, with depth (negated) shown over time. Red lines show the recapture window under consideration and blue lines show any possible moments of recapture in this window.
}
\description{
This function highlights potential recapture events on unknown dates based on a depth threshold; for each day in the depth timeseries, all the timestamps in which the individual was above this depth are considered possible recapture events. These timestamps can be saved by the function and inspected graphically.
}
\examples{

#### Example (1): Suggest recapture events for a single individual
suggested_recap <-
  suggest_recapture(data = dat_flapper[dat_flapper$id == "A", ],
                    threshold_depth = 1,
                    plot = TRUE,
                    window = 60 * 60 * 12,
                    xlab = "Timestamp",
                    ylab = "Depth",
                    prompt = FALSE,
                    ndates_warning = 100,
                    type = "l"
  )
utils::str(suggested_recap)

#### Example (2) Suggest recapture events for multiple individuals by implementing lapply()
dat_flapper_ls <- split(dat_flapper, f = dat_flapper$id)
suggested_recap_ls <-
  lapply(dat_flapper_ls, function(df){
    suggested_recap <-
      suggest_recapture(data = df,
                        threshold_depth = 1,
                        plot = TRUE,
                        window = 60 * 60 * 12,
                        xlab = "Timestamp",
                        ylab = "Depth",
                        prompt = FALSE,
                        ndates_warning = 100,
                        type = "l"
      )
    return(suggested_recap)
  })
utils::str(suggested_recap_ls)

}
\author{
Edward Lavender
}
